pizzaops-intelligence/requirements.txt
streamlit>=1.31.0
pandas>=2.1.0
numpy>=1.24.0
plotly>=5.18.0
scikit-learn>=1.3.0
xgboost>=2.0.0
statsmodels>=0.14.0
prophet>=1.1.5
shap>=0.44.0
openpyxl>=3.1.0
python-pptx>=0.6.21
fpdf2>=2.7.0
holidays>=0.40
anthropic>=0.18.0
python-dotenv>=1.0.0
